requirements.txt
numpy
torch
pyyaml
tqdm
rich
